==> 01_eda.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

sns.set_theme(style="whitegrid")

# Notes

If we use tags as predictors, we will end up with a lot of predictors.
Hence, we need some sort of variable selection method.
We can use lasso, stepwise, etc.

books = pd.read_csv("./goodbooks-10k/books.csv")
book_tags = pd.read_csv("./goodbooks-10k/book_tags.csv")
ratings = pd.read_csv("./goodbooks-10k/ratings.csv")
tags = pd.read_csv("./goodbooks-10k/tags.csv")
to_read = pd.read_csv("goodbooks-10k/to_read.csv")

print(len(ratings.user_id.unique()))

ratings.describe()

len(ratings.user_id.unique())

books.describe()

books.apply(lambda x: len(x.unique()), axis = 0)

ratings.shape

tags.shape

book_tags.head()

(
    book_tags.merge(tags, on = "tag_id")
    .query("tag_id != 30574")
    .query("count > 5000")
    .sort_values("count", ascending=False)
)
     

len(books.work_id.unique())

len(books.goodreads_book_id.unique())

## Tags

- There are 
- Some tags should be ignored
    - ex: currently-reading, to-read
- Some tags are equivalent but spelled differently
    - ex: favorites, favourites, F a v o r i t e s

- Choose p tags as predictors

TAGS_TO_IGNORE = {30574: "to-read", 
                  8717: "currently-reading",
                 }

popular_tags = (
    book_tags.merge(tags, on = "tag_id")
    .query("tag_id not in @TAGS_TO_IGNORE.keys()")
    .query("count > 2000")
    .tag_name
    .unique()
)

# Top raters

ratings

(
    ratings
    .groupby("user_id", as_index=False)
    .count()
    .sort_values("rating", ascending=False)
)

top_user = 30944

## Tags per book

(
    book_tags
    .merge(tags, on="tag_id")
    .groupby(["goodreads_book_id"], as_index=False)
    .count()
    .sort_values("count", ascending=False)
)

# Top tags

top_tags = (
    book_tags
    .merge(books, on="goodreads_book_id")
    .filter(items=["count", "goodreads_book_id", "tag_id", "authors", "ratings_count", "original_title"])
    .merge(tags, on="tag_id")
    .groupby("tag_name", as_index=False)
    .sum()
    .sort_values("count", ascending=False)
    #.query("tag_name != 'to-read'")
    .head(10)
)

top_tags
