==> 02_data_cleantags.ipynb <==
## Goal

There are 34,252 unique tag ids.
To reduce the number of parameters, we will manually select only
the top $n$ most popular tags that meet a certain criteria.
    
Tags should exclusively be metadata about a book and
should not contain information about the user-book relationship.
Popularity will be used to determine the pool of possible tags.
Then, a book will be determined to have a specific tag if 
a set proportion/number of that tags has been reached.
Or the value for a tag is the proportion of books that contained
that tag.

Does a different model need to be created for each person?

A global model where each person is also a predictor would imply that 
the effect of a person is linear.
This is problematic because we would expect each person to have
different tastes in books.
Furthermore, a global model may require a large number of parameters,
at least one for each user.
We can mitigate this by using only the reviews for a set number of users.

Instead we could attempt a polynomial regression to account for the
user-book property interactions.
This could result in a large number of predictors.

1. Remove tags that should not be used:
    1. "owned", "my books", "read", "favorites"

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

books = pd.read_csv("./goodbooks-10k/books.csv")
book_tags = pd.read_csv("./goodbooks-10k/book_tags.csv")
ratings = pd.read_csv("./goodbooks-10k/ratings.csv")
tags = pd.read_csv("./goodbooks-10k/tags.csv")
to_read = pd.read_csv("goodbooks-10k/to_read.csv")

tags.dtypes

sum(tags.tag_name.isna())

len(tags.tag_id.unique())

tags.head()

def filter_words_from_tag(df: pd.DataFrame, bad_tag_words: [str]) -> pd.DataFrame:
    filtered_tags = df.copy()
    for tag in bad_tag_words:
        filtered_tags = filtered_tags[~filtered_tags.tag_name.str.contains(tag)]
    return filtered_tags

def filter_basic_tags(tags: pd.DataFrame) -> pd.DataFrame:
    """
    Filters out unneccessary tags in a basic way.
    """
    return tags[
        ~tags.tag_name.str.contains("read") &
        ~tags.tag_name.str.contains("own") &
        ~tags.tag_name.str.contains("my-") &
        ~tags.tag_name.str.contains("my_") &
        ~(tags.tag_name.str.contains("default"))
    ]

bad_tags = {"read", "own", "my-", "my_", "default", "favorites", "favourites",
           "kindle", "book-club", "library", "audiobook", "audiobooks", "ebook",
           "to-buy", "series", "audio", "novels", "literature"}
tags_filter01 = filter_words_from_tag(tags, bad_tags)

## Most popular tags by total count

def get_most_popular_n_tags(book_tags, tags, n):
    return (
        book_tags.merge(tags, on = "tag_id")
        .groupby(["tag_id", "tag_name"], as_index=False)
        .sum()
        .sort_values("count", ascending=False)
        .head(n=n)
    )

top30_tags_df = get_most_popular_n_tags(book_tags, tags, n=30)
top10_tags_df = get_most_popular_n_tags(book_tags, tags, n=10)
#top10_tags_filter = get_most_popular_n_tags(book_tags, tags_filter01, n=10)

## Save books tags with top tags

### Top 30

(
    book_tags
    .query("tag_id in @top30_tags_df.tag_id")
).to_csv("top30tags/book_tags_01_top30tags.csv", index=False)

top30_tag_names = (
    book_tags.merge(tags, on = "tag_id")
    .groupby(["tag_id", "tag_name"], as_index=False)
    .sum()
    .sort_values("count", ascending=False)
    .head(n=30)
    .tag_name
    .unique()
)

(
    book_tags
    .query("tag_id in @top10_tags_df.tag_id")
    .to_csv("top10tags/book_tags.csv", index=False)
)

## Most popular filter01 tags by total count

(
    book_tags.merge(tags_filter01, on = "tag_id")
    .groupby(["tag_id", "tag_name"], as_index=False)
    .sum()
    .sort_values("count", ascending=False)
    .head(n=30)
    .tag_name
    .unique()
)

## Number of tags per book

book_tags.head()

books.head()

(
    book_tags.merge(tags_filter01, on = "tag_id")
    .merge(books, on = "goodreads_book_id")
    .groupby(["original_title","goodreads_book_id"], as_index=False)
    .sum()
    .sort_values("count", ascending=False)
    .head(n=200)
)

## Which are the genre tags

get_most_popular_n_tags(book_tags, tags_filter01, n=20)

top200tags = get_most_popular_n_tags(book_tags, tags_filter01, n=200)

top200tags[top200tags.tag_name.str.contains("sci")]
top200tags[top200tags.tag_name.str.contains("ya")]
top200tags[top200tags.tag_name.str.contains("classic")]
top200tags[top200tags.tag_name.str.contains("romance")]

tag_identity_map = {
    "sci-fi": "science-fiction",
    "scifi": "science-fiction",
    "ya": "young-adult",
    "classic": "classics",
    "children": "childrens",
    "children-s": "childrens",
    "historical": "historical-fiction",
    "nonfiction": "non-fiction",
    "dystopia": "dystopian",
    "kids": "childrens"
}

tag_id_map = {}
for from_word, to_word in tag_identity_map.items():
    from_id = tags[tags.tag_name == from_word].tag_id.to_numpy()[0]
    to_id = tags[tags.tag_name == to_word].tag_id.to_numpy()[0]
    tag_id_map[from_id] = to_id
tag_id_map

def consolidate_tags(book_tags, tag_id_map) -> pd.DataFrame:
    cleaned_book_tags = book_tags.copy()
    for from_id, to_id in tag_id_map.items():
        cleaned_book_tags.loc[cleaned_book_tags.tag_id == from_id, "tag_id"] = to_id
    return cleaned_book_tags.query("tag_id != 11743")

consolidate_tags(book_tags, tag_id_map).query("tag_id != 11743")

get_most_popular_n_tags(consolidate_tags(book_tags, tag_id_map), tags_filter01, 50)

tag_identity_df = pd.DataFrame({"from": list(tag_identity_map.keys()),
                                "to": list(tag_identity_map.values())})

tag_ids = (
    book_tags
    .merge(tags, on="tag_id")
    .merge(tag_identity_df, left_on = "tag_name", right_on = "from")
    .merge(tags, left_on = "to", right_on = "tag_name")
    .filter(items=["tag_id_x", "tag_id_y"])
    .drop_duplicates()
)

dict(zip(tag_ids.tag_id_x, tag_ids.tag_id_y))

(
    book_tags
    .merge(tags, on="tag_id")
    .merge(tag_identity_df, left_on = "tag_name", right_on = "from")
    .merge(tags, left_on = "to", right_on = "tag_name")
    .filter(items=["tag_id_x", "tag_name_x", "tag_id_y", "tag_name_y"])
    .drop_duplicates()
    #.assign(tag_id_x=lambda x: x.tag_id_y)
)

cleaned_book_tags = (
    book_tags.copy()
    .query("tag_id != 11743")
)
cleaned_book_tags.loc[cleaned_book_tags.tag_id == 32989, "tag_id"] = 33114
cleaned_book_tags.loc[cleaned_book_tags.tag_id == 6857, "tag_id"] = 6953
cleaned_book_tags.loc[cleaned_book_tags.tag_id == 7404, "tag_id"] = 7457
cleaned_book_tags.loc[cleaned_book_tags.tag_id == 26771, "tag_id"] = 26837
cleaned_book_tags.loc[cleaned_book_tags.tag_id == 26894, "tag_id"] = 26837
cleaned_book_tags.loc[cleaned_book_tags.tag_id == 11467, "tag_id"] = 14487

top10_tags_filter = get_most_popular_n_tags(cleaned_book_tags, tags_filter01, n=10)

## Save to csv top 10 filtered tags

#(
#    book_tags
#    .query("tag_id in @top10_tags_filter.tag_id")
#    .to_csv("top10tags/book_tags_filter.csv", index=False)#
#)

(
    cleaned_book_tags
    .merge(tags, on="tag_id")
    .merge(tag_identity_df, left_on = "tag_name", right_on = "from")
    .merge(tags, left_on = "to", right_on = "tag_name")
    #.assign(tag_id_x=lambda x: x.tag_id_y)
)
